# legacy-docs/original-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/artwork/api/new": {
            "post": {
                "summary": "작품 등록",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "department": {
                                        "type": "string"
                                    },
                                    "exhibitionId": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "medium": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "201": {
                            "description": "작품 등록 성공"
                        },
                        "400": {
                            "description": "잘못된 요청"
                        },
                        "401": {
                            "description": "인증 실패"
                        },
                        "500": {
                            "description": "서버 오류"
                        }
                    }
                }
            }
        },
        "/artwork/api/{id}": {
            "put": {
                "summary": "작품 수정",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "year": {
                                        "type": "integer"
                                    },
                                    "medium": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "작품 수정 성공"
                        },
                        "400": {
                            "description": "잘못된 요청"
                        },
                        "404": {
                            "description": "작품을 찾을 수 없음"
                        },
                        "500": {
                            "description": "서버 오류"
                        }
                    }
                }
            },
            "delete": {
                "summary": "작품 삭제",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "작품 삭제 성공"
                    },
                    "404": {
                        "description": "작품을 찾을 수 없음"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/artwork/api/list": {
            "get": {
                "summary": "작품 목록 조회",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    },
                    {
                        "in": "query",
                        "name": "sortField",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "createdAt"
                        }
                    },
                    {
                        "in": "query",
                        "name": "sortOrder",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "in": "query",
                        "name": "keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "exhibition",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "year",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "작품 목록 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "pageInfo": {
                                                    "type": "object",
                                                    "properties": {
                                                        "currentPage": {
                                                            "type": "integer"
                                                        },
                                                        "totalPages": {
                                                            "type": "integer"
                                                        },
                                                        "hasPrev": {
                                                            "type": "boolean"
                                                        },
                                                        "hasNext": {
                                                            "type": "boolean"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/artwork/api/featured": {
            "get": {
                "summary": "추천 작품 목록 조회",
                "responses": {
                    "200": {
                        "description": "추천 작품 목록 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/artwork/api/detail/{slug}": {
            "get": {
                "summary": "작품 상세 정보 조회",
                "parameters": [
                    {
                        "in": "path",
                        "name": "slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "작품 상세 정보 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "작품을 찾을 수 없음"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        },
        "/artwork/api/exhibiting": {
            "post": {
                "summary": "작품을 전시회에 출품",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "artworkId": {
                                        "type": "string"
                                    },
                                    "exhibitionId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "responses": {
                        "200": {
                            "description": "출품 성공"
                        },
                        "400": {
                            "description": "잘못된 요청"
                        },
                        "404": {
                            "description": "작품 또는 전시회를 찾을 수 없음"
                        },
                        "500": {
                            "description": "서버 오류"
                        }
                    }
                }
            }
        },
        "/artwork/api/exhibiting/{artworkId}/{exhibitionId}": {
            "delete": {
                "summary": "작품의 전시회 출품 취소",
                "parameters": [
                    {
                        "in": "path",
                        "name": "artworkId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "exhibitionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "출품 취소 성공"
                    },
                    "404": {
                        "description": "작품 또는 전시회를 찾을 수 없음"
                    },
                    "500": {
                        "description": "서버 오류"
                    }
                }
            }
        }
    }
}